+++ encryption-service/authn/authn.proto
@@ -13,17 +13,35 @@
 // limitations under the License.
 syntax = "proto3";
 
-package authz;
+package authn;
 option go_package = ".;authn";
 
+service Encryptonize{
+  // Creates a new user on the service
+  rpc CreateUser (CreateUserRequest) returns (CreateUserResponse){}
+
+  // User Login
+
+}
+
+enum UserScope{
+  READ = 0;
+  CREATE = 1;
+  INDEX = 2;
+  OBJECTPERMISSIONS = 3;
+  USERMANAGEMENT = 4;
+}
+
+message CreateUserRequest{
+  repeated UserScope user_scopes = 1;
+}
+
+message CreateUserResponse{
+  string user_id = 1;
+  string access_token = 2;
+}
+
 message AccessTokenClient{
-  enum UserScope{
-    READ = 0;
-    CREATE = 1;
-    INDEX = 2;
-    OBJECTPERMISSIONS = 3;
-    USERMANAGEMENT = 4;
-  }
   bytes user_id = 1;
   repeated UserScope user_scopes = 2;
 }
