+++ encryption-service/makefile
@@ -19,8 +19,8 @@ help:  ## Display this help
 
 ##### Files #####
 binary = encryption-service
-protobufs = app/app_grpc.pb.go app/app.pb.go authn/token.pb.go authz/access_object.pb.go
-protosource = app/app.proto authn/token.proto authz/access_object.proto
+protobufs = app/app_grpc.pb.go app/app.pb.go authn/authn_grpc.pb.go authn/authn.pb.go authz/access_object.pb.go
+protosource = app/app.proto authn/authn.proto authz/access_object.proto
 credentials = data/client.root.crt data/object_storage_key data/ca.crt data/client.root.key data/encryptonize.crt data/object_storage_id data/object_storage.crt
 coverage = coverage-unit.html coverage-e2e.html coverage-all.html
 
@@ -32,7 +32,8 @@ build: $(protobufs)  ## Build the Encryption Service binary
 
 $(protobufs): $(protosource)
 	protoc --go-grpc_out=app --go_out=app app/app.proto
-	protoc --go_out=authn authn/token.proto
+	protoc --go-grpc_out=authn --go_out=authn authn/authn.proto
+	protoc --go_out=authn authn/authn.proto
 	protoc --go_out=authz authz/access_object.proto
 
 .PHONY: docker-build
