+++ encryption-service/tests/grpc_e2e_tests/client.go
@@ -22,12 +22,14 @@ import (
 	"google.golang.org/grpc/metadata"
 
 	"encryption-service/app"
+	"encryption-service/authn"
 )
 
 // Client for making test gRPC calls to the encryption service
 type Client struct {
 	connection *grpc.ClientConn
-	client     app.EncryptonizeClient
+	appClient  app.EncryptonizeClient
+	authClient authn.EncryptonizeClient
 	ctx        context.Context
 }
 
@@ -51,13 +53,15 @@ func NewClient(endpoint, token string, https bool) (*Client, error) {
 		return nil, err
 	}
 
-	client := app.NewEncryptonizeClient(connection)
+	appClient := app.NewEncryptonizeClient(connection)
+	authClient := authn.NewEncryptonizeClient(connection)
 	authMetadata := metadata.Pairs("authorization", fmt.Sprintf("bearer %v", token))
 	ctx := metadata.NewOutgoingContext(context.Background(), authMetadata)
 
 	return &Client{
 		connection: connection,
-		client:     client,
+		appClient:  appClient,
+		authClient: authClient,
 		ctx:        ctx,
 	}, nil
 }
@@ -77,7 +81,7 @@ func (c *Client) Store(plaintext, associatedData []byte) (*app.StoreResponse, er
 		},
 	}
 
-	storeResponse, err := c.client.Store(c.ctx, storeRequest)
+	storeResponse, err := c.appClient.Store(c.ctx, storeRequest)
 	if err != nil {
 		return nil, fmt.Errorf("Store failed: %v", err)
 	}
@@ -88,7 +92,7 @@ func (c *Client) Store(plaintext, associatedData []byte) (*app.StoreResponse, er
 func (c *Client) Retrieve(oid string) (*app.RetrieveResponse, error) {
 	retrieveRequest := &app.RetrieveRequest{ObjectId: oid}
 
-	retrieveResponse, err := c.client.Retrieve(c.ctx, retrieveRequest)
+	retrieveResponse, err := c.appClient.Retrieve(c.ctx, retrieveRequest)
 	if err != nil {
 		return nil, fmt.Errorf("Retrieve failed: %v", err)
 	}
@@ -99,7 +103,7 @@ func (c *Client) Retrieve(oid string) (*app.RetrieveResponse, error) {
 func (c *Client) GetPermissions(oid string) (*app.GetPermissionsResponse, error) {
 	getPermissionsRequest := &app.GetPermissionsRequest{ObjectId: oid}
 
-	getPermissionsResponse, err := c.client.GetPermissions(c.ctx, getPermissionsRequest)
+	getPermissionsResponse, err := c.appClient.GetPermissions(c.ctx, getPermissionsRequest)
 	if err != nil {
 		return nil, fmt.Errorf("GetPermissions failed: %v", err)
 	}
@@ -113,7 +117,7 @@ func (c *Client) AddPermission(oid, target string) (*app.AddPermissionResponse, 
 		Target:   target,
 	}
 
-	addPermissionResponse, err := c.client.AddPermission(c.ctx, addPermissionRequest)
+	addPermissionResponse, err := c.appClient.AddPermission(c.ctx, addPermissionRequest)
 	if err != nil {
 		return nil, fmt.Errorf("AddPermission failed: %v", err)
 	}
@@ -127,7 +131,7 @@ func (c *Client) RemovePermission(oid, target string) (*app.RemovePermissionResp
 		Target:   target,
 	}
 
-	removePermissionResponse, err := c.client.RemovePermission(c.ctx, removePermissionRequest)
+	removePermissionResponse, err := c.appClient.RemovePermission(c.ctx, removePermissionRequest)
 	if err != nil {
 		return nil, fmt.Errorf("RemovePermission failed: %v", err)
 	}
@@ -135,12 +139,12 @@ func (c *Client) RemovePermission(oid, target string) (*app.RemovePermissionResp
 }
 
 // Perform a `CreateUser` request.
-func (c *Client) CreateUser(userscopes []app.CreateUserRequest_UserScope) (*app.CreateUserResponse, error) {
-	createUserRequest := &app.CreateUserRequest{
+func (c *Client) CreateUser(userscopes []authn.UserScope) (*authn.CreateUserResponse, error) {
+	createUserRequest := &authn.CreateUserRequest{
 		UserScopes: userscopes,
 	}
 
-	createUserResponse, err := c.client.CreateUser(c.ctx, createUserRequest)
+	createUserResponse, err := c.authClient.CreateUser(c.ctx, createUserRequest)
 	if err != nil {
 		return nil, fmt.Errorf("CreateUser failed: %v", err)
 	}
@@ -149,7 +153,7 @@ func (c *Client) CreateUser(userscopes []app.CreateUserRequest_UserScope) (*app.
 
 // Perform a `Version` request.
 func (c *Client) GetVersion() (*app.VersionResponse, error) {
-	versionResponse, err := c.client.Version(c.ctx, &app.VersionRequest{})
+	versionResponse, err := c.appClient.Version(c.ctx, &app.VersionRequest{})
 
 	if err != nil {
 		return nil, fmt.Errorf("Get version failed: %v", err)
